pyproject.toml
[tool.poetry]
name = "py_social_network"
version = "0.1.0"
description = "Social network by Django"
authors = ["Alexey Kudimov"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.6"
djangorestframework = "^3.12.4"
Pillow = "^8.3.1"
psycopg2 = "^2.9.1"
djoser = "^2.1.0"
drf-yasg = "^1.20.0"
django-cors-headers = "^3.8.0"
django-mptt = "^0.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
